Conflicts:
	UnitTest/Utf8/test_Utf8String.cpp
	Ystring/Generic/GenericString-impl.hpp
	Ystring/Utf8/Utf8String.cpp
	Ystring/Utf8/Utf8String.hpp

+++ Ystring/Generic/GenericString.hpp
@@ -84,7 +84,7 @@ template <typename Str, typename It, typename Enc>
 std::vector<Str> split(
         Range<It> str,
         Enc encoding,
-        int maxParts = 0,
+        ptrdiff_t maxParts = 0,
         SplitFlags_t flags = SplitFlags::IGNORE_EMPTY);
 
 template <typename Str, typename It1, typename It2, typename Enc>
@@ -92,7 +92,7 @@ std::vector<Str> split(
         Range<It1> str,
         Range<It2> cmp,
         Enc encoding,
-        int maxParts = 0,
+        ptrdiff_t maxParts = 0,
         SplitFlags_t flags = SplitFlags::IGNORE_EMPTY);
 
 template <typename Str, typename It, typename Enc, typename Predicate>
@@ -100,7 +100,7 @@ std::vector<Str> splitIf(
         Range<It> str,
         Enc encoding,
         Predicate predicate,
-        int maxParts,
+        ptrdiff_t maxParts,
         SplitFlags_t flags);
 
 template <typename It1, typename It2, typename Enc>
